* Fixed potential endless recursion introduced by SVN r3444   (patch by Ivan Maguidhir). * Bumped version (2.1.0.svn4).
git-svn-id: https://svn.linuxsampler.org/svn/linuxsampler@3452 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ trunk/src/engines/common/Note.h
@@ -34,10 +34,10 @@ namespace LinuxSampler {
 
     // remove strictness of C++ regarding raw bitwise operations (on type release_trigger_t)
     inline release_trigger_t operator|(release_trigger_t a, release_trigger_t b) {
-        return (release_trigger_t) (a | b);
+        return (release_trigger_t) ((int)a | (int)b);
     }
     inline release_trigger_t& operator|=(release_trigger_t& a, release_trigger_t b) {
-        a = (release_trigger_t) (a | b);
+        a = (release_trigger_t) ((int)a | (int)b);
         return a;
     }
